+++ src/uu/users/Cargo.toml
@@ -18,6 +18,9 @@ path = "src/users.rs"
 clap = { workspace = true }
 uucore = { workspace = true, features = ["utmpx"] }
 
+[target.'cfg(target_os = "openbsd")'.dependencies]
+utmp-classic = { workspace = true }
+
 [[bin]]
 name = "users"
 path = "src/main.rs"
